backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.51.2
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
reportlab==4.0.7
Pillow==10.1.0